Added a debug testing mechanism thanks to Matt Neuburg
When I tried to create test methods to verify the code, I needed to set breakpoints deep inside the geometry code. Unfortunately, the tests are performed after the app is started and the display code would be subject to the same breakpoints and would break before my test methods were run. The use of a test phase environment variable makes life much easier. It would be so much better if Xcode simply turned off breakpoints for the target app and turned them back on while the tests were run.